Fix: Negative telemetry values; temp not working

+++ common/ud3core/tasks/tsk_analog.h
@@ -79,7 +79,7 @@ typedef struct
 /* `#END` */
 
 void tsk_analog_Start(void);
-uint16_t CT1_Get_Current(uint8_t channel);
+uint32_t CT1_Get_Current(uint8_t channel);
 float CT1_Get_Current_f(uint8_t channel);
 uint16_t get_max(void);
 void i2t_set_limit(uint32_t const_current, uint32_t ovr_current, uint32_t limit_ms);
